# storage/api-docs/hrm-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini ERP HRM API",
        "description": "Human Resources Management API Documentation for Mini ERP App - Includes Employee Management, Attendance, Leave, Payroll, Reimbursements, Resignations, Office Locations, and Analytics & Reporting",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/platform/hrm/shifts": {
            "get": {
                "tags": [
                    "HRM Shifts"
                ],
                "summary": "List all shifts",
                "operationId": "c8da7dbf3d8962967153e252e484367b",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "HRM Shifts"
                ],
                "summary": "Create a new shift",
                "operationId": "5af93b8340b57ac6d938f8f4f04f033a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Shift created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platform/hrm/shifts/{id}": {
            "get": {
                "tags": [
                    "HRM Shifts"
                ],
                "summary": "Get shift details",
                "operationId": "d52d6eb4dbfb8e98d26935bade42fd66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Shift not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "HRM Shifts"
                ],
                "summary": "Update a shift",
                "operationId": "b1ce87f3e2d3aee745dbf43bb97ac709",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Shift updated"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Shift not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "HRM Shifts"
                ],
                "summary": "Delete a shift",
                "operationId": "f13e56d18c47e25b7dbccd7641fb56d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shift deleted"
                    },
                    "404": {
                        "description": "Shift not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a token from Sanctum auth",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}